feat:se implementa valdiacion de credenciales
