+++ assets/index.php
@@ -1,0 +1,2 @@
+<?php
+	// Silence golden is. May the Force be with you
