+++ readme.txt
@@ -1,0 +1,52 @@
+=== ePayco plugin for WooCommerce ===
+Contributors: ePayco Team
+Donate link: https://epayco.com/
+Tags: e-commerce, commerce, wordpress ecommerce, store, sales, sell, shop, shopping, cart, checkout, configurable, epayco
+Requires at least: 5.5
+Tested up to: 6.4
+Stable tag: 7.0.0
+Requires PHP: 5.6.0
+License: GNU General Public License v3.0
+License URI: http://www.gnu.org/licenses/gpl-3.0.html
+
+Receive payments, accept credit cards, deposits and transfers.
+
+== Description ==
+
+
+ePayco helps you grow your business with rates tailored to your needs.
+Receive payments, accept credit cards, deposits and transfers.
+
+== Installation ==
+
+1. Download the plugin
+2. Enter the administrator of your WordPress.
+3. Enter Plugins / Add-New / Upload-Plugin.
+4. Find the plugin downloaded on your computer and upload it like any other file.
+5. After installing the .zip you can see it in the list of installed plugins, you can activate or deactivate it.
+6. To configure the plugin you must go to: WooCommerce / Settings / Payments and Locate the tab ePayco.
+7. Configure the plugin by entering P_CUST_ID_CLIENTE, P_KEY, PUBLIC_KEY and PRIVATE_KEY provided by ePayco
+8. Save Changes, if you have not done the configuration correctly you will be shown a warning, pay attention to this.
+
+
+== Frequently Asked Questions ==
+
+= What types of payment methods do they receive? =
+
+accept credit cards, debit cars, cash,  deposits and transfers.
+
+= Are you required to use a ssl certificate ? =
+
+Yes. But it is advisable that you consider using it since it is revealing for browsers
+
+== Screenshots ==
+
+1. The front-end view
+
+== Changelog ==
+
+= 6.0.0 =
+* Initial stable release
+
+== Additional Info ==
+**Contribute** [repository on github](https://github.com/epayco/Plugin_ePayco_WooCommerce)
